requirements.txt
requests==2.23.0
pandas==1.0.5
tensorflow==2.4.0
numpy==1.19.5
streamlit==0.79.0
validators==0.18.2
Pillow==8.1.2